New document titling code
This code has a main function that takes in a file path that contains .txt and/or .docx files. It will then generate a dictionary of current titles as the keys and potential new titles as the values. It will iterate through this dictionary and rename all the documents in the given file path to their new names if there is one. 
Yet to be done:
- rename all the pdf's in a separate folder that have the same name. 
- refine the naming scheme to not give as many dud names