Added updated version of documenttitling2 file
This file contains new methods that instead attempt to rename the document using the old name as a reference. Adds relevant date and or WIN number to the document title if they can be found. Old functionality should be able to be obtained by changing the method call of generate_title_dictionary_2(path) to generate_title_dictionary(path) at the end of the main method